src/modules/cars/useCases/car/list-available-cars/ListAvailableCarsUsecase.ts
import { Car } from '@modules/cars/infra/typeorm/entities';
import { ICarsRepository } from '@modules/cars/repositories/ICarsRepository';

interface IRequest {
  brand?: string;
  category_id?: string;
  name?: string;
}

export class ListAvailableCarsUsecase {
  constructor(private carsRepository: ICarsRepository) { }

  async execute({ brand, category_id, name }: IRequest): Promise<Car[]> {
    const cars = await this.carsRepository.findAvailable({
      brand,
      category_id,
      name,
    });
    return cars;
  }
}
